requirements.txt
openai
whisper
sounddevice
soundfile
requests
flask
pytest
pyyaml